separate imap build function
